Add exercise files url to Course object Add download speed throttling Add json reponse examples from pluralsight api Fix file and folder naming

+++ PluralsightDownloader.Web/ViewModel/ProgressArgs.cs
@@ -4,9 +4,9 @@
     {
         public string Id { get; set; }
 
-        public int TotalBytes { get; set; }
+        public long TotalBytes { get; set; }
 
-        public int BytesReceived { get; set; }
+        public long BytesReceived { get; set; }
 
         public float PercentComplete { get { return (float)BytesReceived / TotalBytes; } }
